Added Profile repository

+++ Backend/src/main/java/edu/depaul/cdm/se452/rfa/profileManagement/repository/ProfileRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileRepository extends JpaRepository<Profile, Integer> {
     // You can add custom query methods here, if needed
-    Profile findByUserId(Long userId);
+    Profile findByUserId(Integer userId);
 }
